requirements.txt
# Core dependencies
requests>=2.28.0

# Web UI
flask>=2.3.0

# Environment variable loading
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# Machine learning (recommendations)
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
